backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
faiss-cpu==1.8.0.post1
numpy==2.1.2
apify-client==1.8.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
httpx==0.27.2
